Improve DecimalField
This fixes a major issue with the old implementation, in that it would not pick up the changed value if the user hit a "Done" button while the field was in focus and presented in a sheet (I didn't check if the bug exists in an "inline" form).

The key is the `.onReceive` modifier that receives a stream of values as the user types  into the field and lets us process those values.

It also changes the behaviour somewhat, which might be a deal-breaker for you or some people: the field now shows the formatted value (ie. `$10.00`) like before, but as soon as the user focuses input on the field it changes back to a plain value (ie. `10`) to ease editing. On blur, it obviously shows the formatted value again.